Update and move image function to img.go
